テーブルのオブジェクト修正
