+++ src/Model/Table/CommentsTable.php
@@ -99,50 +99,4 @@ class CommentsTable extends Table
         $rules->add($rules->existsIn(['actor_id'], 'Actors'));
         return $rules;
     }
-
-    /**
-     * doReaction method
-     *
-     * comment_reactionsで comment_reactions.actor_idが自分自身でかつ、
-     * comment_reactions.type=$reactionsのものが既にDB内にあれば何もしない
-     * そうでなければ、comment_reactions を下記の条件で作成する
-     * comment_reactions.actor_id = $actor->id
-     * comment_reactions.type=$reactions
-    */
-    public function doReaction($actor, $reactions, $comment_id)
-    //actorはUserのインスタンスとする、comment_idも入れています
-    {
-        // DBに既にあれば登録しない
-        $Comment_Reactions = TableRegistry::get('Comment_Reactions')->find();
-        $result = $Comment_Reactions->where(['comment_id'=>$comment_id, 'actor_id' => $actor->id, 'kind' => $reactions])->first();
-
-        if ($result == null) {
-            $Comment_Reactions = TableRegistry::get('Comment_Reactions');
-            $new_reaction = $Comment_Reactions->newEntity();
-            $data = [
-                'comment_id' => $comment_id,
-                'actor_id' => $actor->id,
-                'kind' => $reactions
-            ];
-            $new_reaction->set($data);
-            $Comment_Reactions->save($new_reaction);
-        }
-    }
-
-    public function good_question($actor)
-    {
-        return doReaction($actor, Reactions::$GOOD_QUESTION);
-    }
-
-    public function nice_advise($actor)
-    {
-        return doReaction($actor, Reactions::$NICE_ADVISE);
-    }
-
-    public function answer($actor)
-    {
-        return doReaction($actor, Reactions::$ANSWER);
-
-    }
-
 }
